RAG_project/requirements.txt
streamlit 
python-dotenv 
PyPDF2
langchain
langchain-community
langchain-text-splitters
openai 
langchain-openai


# source tt/bin/activate 
# tt\Scripts\activate  
# streamlit run


ImportError: Could not import faiss python package. Please install it with `pip install faiss-gpu` (for CUDA supported GPU) or `pip install faiss-cpu` (depending on Python version).






